Refactored `wavecal1d` to use a new `WavelengthSolution1D` class for managing wavelength models and transformations.
